refactor(services): move logic to index files
