+++ src/test-utils/stubs/field.ts
@@ -1,4 +1,4 @@
-import { Factory } from '../../game/services/Factory/Factory'
+import { Factory } from '../../game/services/Factory'
 import { IField } from '../../game/types'
 
 export const stubField = (overrides: Partial<IField> = {}): IField => {
